requirements.txt
requests~=2.31.0
Pillow~=9.5.0
colourpycker~=0.0.6
colorgram.py
